% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ICODS.R
\docType{methods}
\name{ICODS-internal-api}
\alias{ICODS-internal-api}
\alias{tmp}
\alias{minAIC,ICODS-method}
\alias{optimObj,ICODS-method}
\alias{mVal,ICODS-method}
\alias{estimate,ICODS-method}
\alias{print,ICODS-method}
\alias{show,ICODS-method}
\alias{summary,ICODS-method}
\title{Hidden methods}
\usage{
tmp(x)

\S4method{minAIC}{ICODS}(object, ...)

\S4method{optimObj}{ICODS}(object, ...)

\S4method{mVal}{ICODS}(object, ...)

\S4method{estimate}{ICODS}(object, ...)

\S4method{print}{ICODS}(x, ...)

\S4method{show}{ICODS}(object)

\S4method{summary}{ICODS}(object, ...)
}
\description{
Hidden methods
}
\keyword{internal}
